Moved analyzer options to Core
